+++ framework/Source/CPTXYPlotSpace.m
@@ -12,10 +12,8 @@
 #import "CPTPlotArea.h"
 #import "CPTPlotAreaFrame.h"
 #import "CPTUtilities.h"
+#import "NSCoderExtensions.h"
 #import <tgmath.h>
-
-static const CGFloat kCPTMomentumTime = CPTFloat(0.25); // Deceleration time in seconds for momentum scrolling
-static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in seconds when scrolled past the global range
 
 /// @cond
 @interface CPTXYPlotSpace()
@@ -31,8 +29,8 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
 -(double)doublePrecisionPlotCoordinateForViewLength:(CGFloat)viewLength logPlotRange:(CPTPlotRange *)range boundsLength:(CGFloat)boundsLength;
 
 -(CPTPlotRange *)constrainRange:(CPTPlotRange *)existingRange toGlobalRange:(CPTPlotRange *)globalRange;
--(void)animateRange:(CPTPlotRange *)oldRange property:(NSString *)property globalRange:(CPTPlotRange *)globalRange shift:(NSDecimal)shift;
--(CPTPlotRange *)shiftRange:(CPTPlotRange *)oldRange by:(NSDecimal)shift inGlobalRange:(CPTPlotRange *)globalRange elastic:(BOOL)elastic withDisplacement:(CGFloat *)displacement;
+-(void)animateRangeForCoordinate:(CPTCoordinate)coordinate shift:(NSDecimal)shift momentumTime:(CGFloat)momentumTime;
+-(CPTPlotRange *)shiftRange:(CPTPlotRange *)oldRange by:(NSDecimal)shift inGlobalRange:(CPTPlotRange *)globalRange withDisplacement:(CGFloat *)displacement;
 
 @property (nonatomic, readwrite) BOOL isDragging;
 @property (nonatomic, readwrite) CGPoint lastDragPoint;
@@ -119,17 +117,25 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
  **/
 @synthesize allowsMomentum;
 
-/** @property BOOL elasticGlobalXRange
- *  @brief If @YES, the plot space can scroll beyond the bounds set by the @ref globalXRange,
- *  and will bounce back to the @ref globalXRange when released. Defaults to @NO.
+/** @property CPTAnimationCurve momentumAnimationCurve
+ *  @brief The animation curve used to stop the motion of the plot ranges when scrolling with momentum. Defaults to #CPTAnimationCurveQuadraticOut.
  **/
-@synthesize elasticGlobalXRange;
+@synthesize momentumAnimationCurve;
 
-/** @property BOOL elasticGlobalYRange
- *  @brief If @YES, the plot space can scroll beyond the bounds set by the @ref globalYRange,
- *  and will bounce back to the @ref globalYRange when released. Defaults to @NO.
+/** @property CPTAnimationCurve bounceAnimationCurve
+ *  @brief The animation curve used to return the plot range back to the global range after scrolling. Defaults to #CPTAnimationCurveQuadraticOut.
  **/
-@synthesize elasticGlobalYRange;
+@synthesize bounceAnimationCurve;
+
+/** @property CGFloat momentumAcceleration
+ *  @brief Deceleration in pixels/second^2 for momentum scrolling. Defaults to @num{2000.0}.
+ **/
+@synthesize momentumAcceleration;
+
+/** @property CGFloat bounceAcceleration
+ *  @brief Bounce-back acceleration in pixels/second^2 when scrolled past the global range. Defaults to @num{3000.0}.
+ **/
+@synthesize bounceAcceleration;
 
 @dynamic isDragging;
 @synthesize lastDragPoint;
@@ -154,8 +160,10 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
  *  - @ref xScaleType = #CPTScaleTypeLinear
  *  - @ref yScaleType = #CPTScaleTypeLinear
  *  - @ref allowsMomentum = @NO
- *  - @ref elasticGlobalXRange = @NO
- *  - @ref elasticGlobalYRange = @NO
+ *  - @ref momentumAnimationCurve = #CPTAnimationCurveQuadraticOut
+ *  - @ref bounceAnimationCurve = #CPTAnimationCurveQuadraticOut
+ *  - @ref momentumAcceleration = @num{2000.0}
+ *  - @ref bounceAcceleration = @num{3000.0}
  *
  *  @return The initialized object.
  **/
@@ -174,9 +182,11 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         lastDeltaTime    = 0.0;
         animations       = [[NSMutableArray alloc] init];
 
-        allowsMomentum      = NO;
-        elasticGlobalXRange = NO;
-        elasticGlobalYRange = NO;
+        allowsMomentum         = NO;
+        momentumAnimationCurve = CPTAnimationCurveQuadraticOut;
+        bounceAnimationCurve   = CPTAnimationCurveQuadraticOut;
+        momentumAcceleration   = 2000.0;
+        bounceAcceleration     = 3000.0;
     }
     return self;
 }
@@ -199,8 +209,10 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     [coder encodeInteger:self.xScaleType forKey:@"CPTXYPlotSpace.xScaleType"];
     [coder encodeInteger:self.yScaleType forKey:@"CPTXYPlotSpace.yScaleType"];
     [coder encodeBool:self.allowsMomentum forKey:@"CPTXYPlotSpace.allowsMomentum"];
-    [coder encodeBool:self.elasticGlobalXRange forKey:@"CPTXYPlotSpace.elasticGlobalXRange"];
-    [coder encodeBool:self.elasticGlobalYRange forKey:@"CPTXYPlotSpace.elasticGlobalYRange"];
+    [coder encodeInt:self.momentumAnimationCurve forKey:@"CPTXYPlotSpace.momentumAnimationCurve"];
+    [coder encodeInt:self.bounceAnimationCurve forKey:@"CPTXYPlotSpace.bounceAnimationCurve"];
+    [coder encodeCGFloat:self.momentumAcceleration forKey:@"CPTXYPlotSpace.momentumAcceleration"];
+    [coder encodeCGFloat:self.bounceAcceleration forKey:@"CPTXYPlotSpace.bounceAcceleration"];
 
     // No need to archive these properties:
     // lastDragPoint
@@ -220,9 +232,11 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         xScaleType   = (CPTScaleType)[coder decodeIntegerForKey : @"CPTXYPlotSpace.xScaleType"];
         yScaleType   = (CPTScaleType)[coder decodeIntegerForKey : @"CPTXYPlotSpace.yScaleType"];
 
-        allowsMomentum      = [coder decodeBoolForKey:@"CPTXYPlotSpace.allowsMomentum"];
-        elasticGlobalXRange = [coder decodeBoolForKey:@"CPTXYPlotSpace.elasticGlobalXRange"];
-        elasticGlobalYRange = [coder decodeBoolForKey:@"CPTXYPlotSpace.elasticGlobalYRange"];
+        allowsMomentum         = [coder decodeBoolForKey:@"CPTXYPlotSpace.allowsMomentum"];
+        momentumAnimationCurve = (CPTAnimationCurve)[coder decodeIntForKey : @"CPTXYPlotSpace.momentumAnimationCurve"];
+        bounceAnimationCurve   = (CPTAnimationCurve)[coder decodeIntForKey : @"CPTXYPlotSpace.bounceAnimationCurve"];
+        momentumAcceleration   = [coder decodeCGFloatForKey:@"CPTXYPlotSpace.momentumAcceleration"];
+        bounceAcceleration     = [coder decodeCGFloatForKey:@"CPTXYPlotSpace.bounceAcceleration"];
 
         lastDragPoint    = CGPointZero;
         lastDisplacement = CGPointZero;
@@ -323,7 +337,7 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     if ( ![range isEqualToRange:xRange] ) {
         CPTPlotRange *constrainedRange;
 
-        if ( self.elasticGlobalXRange ) {
+        if ( self.allowsMomentum ) {
             constrainedRange = range;
         }
         else {
@@ -387,7 +401,7 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     if ( ![range isEqualToRange:yRange] ) {
         CPTPlotRange *constrainedRange;
 
-        if ( self.elasticGlobalYRange ) {
+        if ( self.allowsMomentum ) {
             constrainedRange = range;
         }
         else {
@@ -464,10 +478,31 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     }
 }
 
--(void)animateRange:(CPTPlotRange *)oldRange property:(NSString *)property globalRange:(CPTPlotRange *)globalRange shift:(NSDecimal)shift
+-(void)animateRangeForCoordinate:(CPTCoordinate)coordinate shift:(NSDecimal)shift momentumTime:(CGFloat)momentumTime
 {
     NSMutableArray *animationArray = self.animations;
     CPTAnimationOperation *op;
+
+    NSString *property        = nil;
+    CPTPlotRange *oldRange    = nil;
+    CPTPlotRange *globalRange = nil;
+
+    switch ( coordinate ) {
+        case CPTCoordinateX:
+            property    = @"xRange";
+            oldRange    = self.xRange;
+            globalRange = self.globalXRange;
+            break;
+
+        case CPTCoordinateY:
+            property    = @"yRange";
+            oldRange    = self.yRange;
+            globalRange = self.globalYRange;
+            break;
+
+        default:
+            break;
+    }
 
     CPTMutablePlotRange *newRange = [oldRange mutableCopy];
 
@@ -480,9 +515,9 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
                           property:property
                      fromPlotRange:oldRange
                        toPlotRange:newRange
-                          duration:kCPTMomentumTime
-                    animationCurve:CPTAnimationCurveQuadraticOut
-                          delegate:nil];
+                          duration:momentumTime
+                    animationCurve:self.momentumAnimationCurve
+                          delegate:self];
         [animationArray addObject:op];
     }
 
@@ -490,14 +525,43 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         CPTPlotRange *constrainedRange = [self constrainRange:newRange toGlobalRange:globalRange];
 
         if ( ![newRange isEqualToRange:constrainedRange] ) {
+            CPTCoordinate orthogonalCoordinate = CPTOrthogonalCoordinate(coordinate);
+
+            NSDecimal newPoint[2];
+            newPoint[coordinate]           = newRange.location;
+            newPoint[orthogonalCoordinate] = CPTDecimalFromInteger(1);
+
+            NSDecimal constrainedPoint[2];
+            constrainedPoint[coordinate]           = constrainedRange.location;
+            constrainedPoint[orthogonalCoordinate] = CPTDecimalFromInteger(1);
+
+            CGPoint newViewPoint         = [self plotAreaViewPointForPlotPoint:newPoint numberOfCoordinates:2];
+            CGPoint constrainedViewPoint = [self plotAreaViewPointForPlotPoint:constrainedPoint numberOfCoordinates:2];
+
+            CGFloat offset = CPTFloat(0.0);
+            switch ( coordinate ) {
+                case CPTCoordinateX:
+                    offset = constrainedViewPoint.x - newViewPoint.x;
+                    break;
+
+                case CPTCoordinateY:
+                    offset = constrainedViewPoint.y - newViewPoint.y;
+                    break;
+
+                default:
+                    break;
+            }
+
+            CGFloat bounceTime = sqrt(ABS(offset) / self.bounceAcceleration);
+
             op = [CPTAnimation animate:self
                               property:property
                          fromPlotRange:newRange
                            toPlotRange:constrainedRange
-                              duration:kCPTBounceTime
-                             withDelay:(hasShift ? kCPTMomentumTime : 0.0)
-                        animationCurve:CPTAnimationCurveElasticOut
-                              delegate:nil];
+                              duration:bounceTime
+                             withDelay:(hasShift ? momentumTime : 0.0)
+                        animationCurve:self.bounceAnimationCurve
+                              delegate:self];
             [animationArray addObject:op];
         }
     }
@@ -994,16 +1058,13 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     CPTPlotRange *newRangeX = [[CPTPlotRange alloc] initWithLocation:newLocationX length:newLengthX];
     CPTPlotRange *newRangeY = [[CPTPlotRange alloc] initWithLocation:newLocationY length:newLengthY];
 
-    BOOL oldElasticGlobalXRange = self.elasticGlobalXRange;
-    BOOL oldElasticGlobalYRange = self.elasticGlobalYRange;
-    self.elasticGlobalXRange = NO;
-    self.elasticGlobalYRange = NO;
+    BOOL oldMomentum = self.allowsMomentum;
+    self.allowsMomentum = NO;
 
     self.xRange = newRangeX;
     self.yRange = newRangeY;
 
-    self.elasticGlobalXRange = oldElasticGlobalXRange;
-    self.elasticGlobalYRange = oldElasticGlobalYRange;
+    self.allowsMomentum = oldMomentum;
 }
 
 /// @endcond
@@ -1058,9 +1119,11 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         self.isDragging       = YES;
 
         // Clear any previous animations
-        for ( CPTAnimationOperation *op in self.animations ) {
+        NSMutableArray *animationArray = self.animations;
+        for ( CPTAnimationOperation *op in animationArray ) {
             [[CPTAnimation sharedInstance] removeAnimationOperation:op];
         }
+        [animationArray removeAllObjects];
 
         return YES;
     }
@@ -1111,9 +1174,10 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
                 CGPoint pointInPlotArea = [theGraph convertPoint:interactionPoint toLayer:plotArea];
                 CGPoint displacement    = self.lastDisplacement;
 
+                CGFloat acceleration     = self.momentumAcceleration;
                 CGFloat speed            = sqrt(displacement.x * displacement.x + displacement.y * displacement.y) / CPTFloat(lastDeltaT);
-                CGFloat acceleration     = speed / kCPTMomentumTime;
-                CGFloat distanceTraveled = speed * kCPTMomentumTime - CPTFloat(0.5) * acceleration * kCPTMomentumTime * kCPTMomentumTime;
+                CGFloat momentumTime     = speed / (CPTFloat(2.0) * acceleration);
+                CGFloat distanceTraveled = speed * momentumTime - CPTFloat(0.5) * acceleration * momentumTime * momentumTime;
                 distanceTraveled = MAX( distanceTraveled, CPTFloat(0.0) );
                 CGFloat theta = atan2(displacement.y, displacement.x);
 
@@ -1123,16 +1187,16 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
 
                 // X range
                 NSDecimal shiftX = CPTDecimalSubtract(lastPoint[CPTCoordinateX], newPoint[CPTCoordinateX]);
-                [self animateRange:self.xRange property:@"xRange" globalRange:self.globalXRange shift:shiftX];
+                [self animateRangeForCoordinate:CPTCoordinateX shift:shiftX momentumTime:momentumTime];
 
                 // Y range
                 NSDecimal shiftY = CPTDecimalSubtract(lastPoint[CPTCoordinateY], newPoint[CPTCoordinateY]);
-                [self animateRange:self.yRange property:@"yRange" globalRange:self.globalYRange shift:shiftY];
+                [self animateRangeForCoordinate:CPTCoordinateY shift:shiftY momentumTime:momentumTime];
             }
         }
         else {
-            [self animateRange:self.xRange property:@"xRange" globalRange:self.globalXRange shift:CPTDecimalFromInteger(0)];
-            [self animateRange:self.yRange property:@"yRange" globalRange:self.globalYRange shift:CPTDecimalFromInteger(0)];
+            [self animateRangeForCoordinate:CPTCoordinateX shift:CPTDecimalFromInteger(0) momentumTime:CPTFloat(0.0)];
+            [self animateRangeForCoordinate:CPTCoordinateY shift:CPTDecimalFromInteger(0) momentumTime:CPTFloat(0.0)];
         }
 
         return YES;
@@ -1198,7 +1262,6 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         CPTPlotRange *newRangeX = [self shiftRange:self.xRange
                                                  by:shiftX
                                       inGlobalRange:self.globalXRange
-                                            elastic:self.elasticGlobalXRange
                                    withDisplacement:&displacement.x];
 
         // Y range
@@ -1206,7 +1269,6 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
         CPTPlotRange *newRangeY = [self shiftRange:self.yRange
                                                  by:shiftY
                                       inGlobalRange:self.globalYRange
-                                            elastic:self.elasticGlobalYRange
                                    withDisplacement:&displacement.y];
 
         self.lastDragPoint    = pointInPlotArea;
@@ -1225,7 +1287,7 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     return NO;
 }
 
--(CPTPlotRange *)shiftRange:(CPTPlotRange *)oldRange by:(NSDecimal)shift inGlobalRange:(CPTPlotRange *)globalRange elastic:(BOOL)elastic withDisplacement:(CGFloat *)displacement
+-(CPTPlotRange *)shiftRange:(CPTPlotRange *)oldRange by:(NSDecimal)shift inGlobalRange:(CPTPlotRange *)globalRange withDisplacement:(CGFloat *)displacement
 {
     CPTMutablePlotRange *newRange = [oldRange mutableCopy];
 
@@ -1234,7 +1296,7 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
     if ( globalRange ) {
         CPTPlotRange *constrainedRange = [self constrainRange:newRange toGlobalRange:globalRange];
 
-        if ( elastic ) {
+        if ( self.allowsMomentum ) {
             if ( ![newRange isEqualToRange:constrainedRange] ) {
                 // reduce the shift as we get farther outside the global range
                 NSDecimal rangeLength = newRange.length;
@@ -1258,5 +1320,17 @@ static const CGFloat kCPTBounceTime   = CPTFloat(0.5);  // Bounce-back time in s
 }
 
 /// @}
+
+#pragma mark -
+#pragma mark Animation Delegate
+
+/// @cond
+
+-(void)animationDidFinish:(CPTAnimationOperation *)operation
+{
+    [self.animations removeObjectIdenticalTo:operation];
+}
+
+/// @endcond
 
 @end